refactor: 适配数据库表结构变更，完善文章版本控制机制
- 移除article_detail表的latest和published字段相关代码
- 新增article表的publishedVersion字段支持
- 统一版本号数据类型为Integer
- 删除editToken相关冗余字段
- 完善新文章和新内容的版本初始化逻辑
- 更新相关实体类、DTO、转换器和服务层代码

+++ bytelogs-service/src/main/java/top/harrylei/forum/service/article/repository/dao/ArticleDetailDAO.java
@@ -9,11 +9,13 @@ import top.harrylei.forum.service.article.repository.mapper.ArticleDetailMapper;
 
 /**
  * 文章详细访问对象
+ *
+ * @author harry
  */
 @Repository
 public class ArticleDetailDAO extends ServiceImpl<ArticleDetailMapper, ArticleDetailDO> {
 
-    public int updateArticleContent(Long articleId, String content, Long version) {
+    public int updateArticleContent(Long articleId, String content, Integer version) {
         return getBaseMapper().updateArticleContent(articleId, content, version);
     }
